chore(renderer): add tests for name condition trigger extractor
