+++ packages/react-form-renderer/src/get-condition-triggers/get-condition-triggers.d.ts
@@ -1,0 +1,5 @@
+import { ConditionDefinition } from "../../condition";
+
+declare function getConditionTriggers(params:ConditionDefinition | ConditionDefinition[]): string[];
+
+export default getConditionTriggers;
