+++ packages/react-form-renderer/src/get-condition-triggers/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './get-condition-triggers';
